Time: 0 ms (100.00%) | Memory: 8.7 MB (35.98%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int duplicate=x;
-        long long ans=0;
+        long ans=0;
         while(x>0){
             int last_digit=x%10;
             ans=ans*10+last_digit;
